page_alloc.c

+++ include/misc.h
@@ -16,9 +16,9 @@
 #define PAGE_SHIFT	12
 #define PAGE_SIZE	(1UL << PAGE_SHIFT)
 
-#define PFN_UP(x)	(((x) + PAGE_SIZE-1) >> PAGE_SHIFT)
-#define PFN_DOWN(x)	((x) >> PAGE_SHIFT)
-#define PFN_PHYS(x)	((x) << PAGE_SHIFT)
+#define PFN_UP(x)	(((unsigned long)(x) + PAGE_SIZE-1) >> PAGE_SHIFT)
+#define PFN_DOWN(x)	((unsigned long)(x) >> PAGE_SHIFT)
+#define PFN_PHYS(x)	((unsigned long)(x) << PAGE_SHIFT)
 
 #define SYSCALL_VECTOR	0x80
